+++ ejercicios/app.js
@@ -4,8 +4,7 @@ var cookieParser = require("cookie-parser");
 var logger = require("morgan");
 
 var indexRouter = require("./routes/index");
-var usersRouter = require("./routes/users");
-var postsRouter = require("./routes/users");
+var postsRouter = require("./routes/posts");
 
 var app = express();
 
@@ -16,7 +15,7 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
 
 app.use("/", indexRouter);
-app.use("/users", usersRouter);
-app.use("/posts", usersRouter);
+
+app.use("/posts", postsRouter);
 
 module.exports = app;
