--- ejercicios/routes/addToArray.js
@@ -1,8 +1,0 @@
-const posts = require("./posts");
-
-let addPostToArray = (post) => {
-  posts.push(post);
-  console.log(posts);
-};
-
-module.exports = addPostToArray;
