+++ ejercicios/routes/postsList.js
@@ -1,0 +1,19 @@
+const posts = [
+  {
+    id: 3,
+    user: "Sara",
+    post: "I'm on holiday",
+  },
+  {
+    id: 4,
+    user: "Claire",
+    post: "I have passed my exam",
+  },
+  {
+    id: 5,
+    user: "Emma",
+    post: "I have started working at McDonalds",
+  },
+];
+
+module.exports = posts;
